Bugfixes
* Tiling by window incorrectly threw away data when data to be combined
had NAs for certain sites in a region. This has been fixed by not
having NAs in the methylSigData object. To the best of my knowledge, 0
and NA are equivalent, and this should not affect any calculations.
* Modified test data to include a CpG at the end of chr21.
* Changed tiling by window method to truncate the last region so that
it doesn't go over the chromosome bound. This is done in a genome
agnostic way by letting the region go 2bp after the last CpG (beginning
of last region).
* There is a new tiling method that allows users to tile / aggregate
over annotations represented as a BED3 data.frame.
* Fixed an error in the vignette where minCount = 0 and maxCount=Inf.
Don't think it liked the Inf.
* Updated documentation.
* Updated unit tests to ensure methylSigReadData and methylSigTile
behave as want them to in all the strange cases.