WIP: Fix methylSigReadData errors
The call to bsseq::BSseq after filtering for SNPs and min/maxCount fails with error:

Error in validObject(.Object) : 
  invalid class “SummarizedExperiment” object: 
  nb of rows in 'assay' (4) must equal nb of rows in 'rowData' (5)

I think this is a bug in SummarizedExperiment. I would expect with rmZeroCov = TRUE, that it would do the subsetting for me. However, changing to rmZeroCov = FALSE and doing it manually works fine.